feat(web-iconics):  update folder structure and reduce code lines
