+++ Web/starter_project_iconics/components/common/blogCard.tsx
@@ -1,12 +1,6 @@
 import React from 'react'
 import { Blog } from '../../types/profile/blog'
-import { MdOutlineModeComment } from 'react-icons/md'
-
-import {
-  AiFillClockCircle,
-  AiFillCheckCircle,
-  AiFillWarning,
-} from 'react-icons/ai'
+import BlogFooter from './blogFooter'
 
 interface Props {
   card: Blog
@@ -14,18 +8,6 @@ interface Props {
 }
 
 const BlogCard: React.FC<Props> = ({ card, feature }) => {
-  const statusIcon = {
-    approved: <AiFillCheckCircle size={18} />,
-    declined: <AiFillWarning size={18} />,
-    pending: <AiFillClockCircle size={18} />,
-  }
-
-  const abbreviations: [number, string][] = [
-    [1e3, 'K'],
-    [1e6, 'M'],
-    [1e9, 'B'],
-  ]
-
   return (
     <div className="card w-[400px] font-montserrat flex flex-col border border-gray-300 rounded-lg overflow-hidden shadow-md bg-white h-543 my-8 font-bold">
       <img
@@ -79,77 +61,7 @@ const BlogCard: React.FC<Props> = ({ card, feature }) => {
         </p>
         <br />
         <hr className="line border-gray-200" />
-        <div>
-          <div className="card_footer flex justify-between items-center">
-            {(() => {
-              if (feature === 'likes') {
-                let val: string = ''
-                for (const [value, abbreviation] of abbreviations) {
-                  if (card.likes >= value) {
-                    const roundedNum = (card.likes / value).toFixed(1)
-                    val = `${roundedNum}${abbreviation} likes`
-                  }
-                }
-                return (
-                  <span className="flex items-center">
-                    <span
-                      className="card_writter_likes m-0 text-sm font-medium text-gray-500"
-                      data-likes={card.likes}
-                      style={{ display: 'inline-block', marginRight: '14px' }}
-                    >
-                      <MdOutlineModeComment size={15} />
-                    </span>
-                    <span
-                      className="card_writter_likes m-0 text-sm font-medium text-gray-500"
-                      data-likes={card.likes}
-                      style={{ display: 'inline-block' }}
-                    >
-                      {val}
-                    </span>
-                  </span>
-                )
-              } else {
-                return (
-                  <span className="flex items-center">
-                    {
-                      <span
-                        className={
-                          card.status === 'approved'
-                            ? 'text-green-600 '
-                            : card.status === 'declined'
-                            ? 'text-red-600 '
-                            : 'text-orange-500 '
-                        }
-                        style={{
-                          display: 'inline-block',
-                          marginRight: '10px',
-                        }}
-                      >
-                        {statusIcon[card.status || 'pending']}
-                      </span>
-                    }
-                    <span
-                      className={
-                        card.status === 'approved'
-                          ? 'text-green-600 text-sm'
-                          : card.status === 'declined'
-                          ? 'text-red-600 text-sm'
-                          : 'text-orange-500 text-sm'
-                      }
-                      style={{ display: 'inline-block' }}
-                    >
-                      {card.status.toUpperCase()}
-                    </span>
-                  </span>
-                )
-              }
-            })()}
-
-            <button className="card_read_more px-4 py-2 text-sm rounded-md border-none bg-white text-purple-500 font-bold cursor-pointer transition duration-200 ease-in-out">
-              Read More
-            </button>
-          </div>
-        </div>
+        <BlogFooter card={card} feature={feature} />
       </div>
     </div>
   )
